Make Dates POSIXct when index(prices) is not
The Dates object is created from the index of the Prices argument (or
the object containing data for the given Symbol) if the user doesn't
supply the Dates argument in the updatePortf call.

We subtract a very small value from the first observation in Dates in
order to get a date range for which position P&L needs to be updated.
If the index of Prices is a 'Date' vector, subtracting a very small
value from Dates causes the startDate to be shifted back an entire day.
This causes the initializing transaction in the txn table to be
included in the P&L calculations.

Avoid this by ensuring Dates (and therefore startDate and endDate) are
always POSIXct with the same timezone as the txn table (the index of
the posPL and txn tables are always POSIXct).


git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/blotter/pkg/blotter@1704 edb9625f-4e0d-4859-8d74-9fd3b1da38cb