ADDED DURATION COLUMN TO PER TRADE STATS OUTPUT
Adding a DURATION column will help simulating trades in the blotter::mcsim
function whilst maintaining strategy-specific characteristics

DURATION is simply the difference between the Start and End POSIXct
objects and it return value is a difftime object which includes the units
of the return value. This detail will be key for simulation purposes.