First draft of txnsim
This is a first version of txnsim, which simulates blotter transactions
whilst retaining the original strategy characteristics (namely duration
and quantity). This first version handles trades with
tradeDef="flat.to.flat". Subsequent versions will handle
"flat.to.reduced" and will include additional methods for plotting.

Thanks to Joshua Ulrich for his assistance with PR amendment and general
git-spertise and for his recommendation of "lintr" and style patch.