add storage of .Random.seed for later reproducinbility, if required
The Monte Carlo simulation function in blotter make use of random sampling to create the replicates.  Normally, the user should want these samples to be as random as possible, and should avoid use of set.seed().  It is also true that the ability to reproduce a specific simulation may also be desirable.  As such, .Random.seed state at the beginning of function execution is extracted and stored so that the simulation should be able to be duplicated precisely if required.