fix last open trade and include fees
Using only Period.Realized.PL does not take fees into account.

For flat.to.flat, we can go back to using the sum of Net.Trading.PL.  for flat.to.reduced and increased.to.reduced, we need to in clude only the appropriate fees.  for flat.to.reduced, this is the proportion of fees we can attribute to the closing transaction.  For increased.to.reduced, it is the fees for the opening trade plus the fees for the closing trade.

For the last closing trade, the quantity should be set to the open position, P&L is the sum of the unrealized P&L's (fees not included here because we're not realizing any P&L).

Also, rename Pos.PL to trade.PL to avoid confusion with posPL slot in portfolio.

see #46