make unrealized P&L vary based on proportion of total position
see #47

For flat.to.flat, we know precisely how much of the cumulative P&L to associate with this round turn: all of it.

For flat.to.reduced and increased.to.reduced/acfifo, we need to take a proportion of unrealized P&L equal to the proportional quantity.

For flat.to.reduced, this is not precise, because all trades start at flat, and we can't precisely allocate P&L if the total position is leveled into.  We could consider using proportion of max position as I did originally, but this will undercount gains/losses in many cases.  The current method may overcount, so is more conservative.

For increased.to.reduced/acfifo, we know precisely what the realized P&L and fees are, and only need proportional unrealized P&L, as pointed out by Jasden Mackie.  to get the unrealized P&L, we take the proportion of total position attributed to this round turn and use that as a time-arying pro-rata.

Still outstanding: calculating the correct proportion for an open trade at the end of the series.