Ensure transactions are added in order
The portfolio accounting could be wrong if transactions are added out
of order. Given that this is a strict requirement for correctness, add
a paragraph to the Note section of ?addTxn, and throw an error if
addTxn or addTxns are called with a date before the last date in the
portfolio.