Warn if user passes invalid Dates argument
The previous code checked to ensure the starting date-time in the Dates
ISO8601 string parsed correctly and was within bounds of the data, but
did not do the same for the ending date-time.

Parse the string once, then check the starting and ending date-times
are not NA and are within the bounds of the price data. Warn the user
if either of the bounds are invalid or outside the price data range.
The warning also reports the date range being used.

Fixes #20.