Fix txn-time order check on objects w/Date index
Transactions should always be added to the portfolio in time-order.
Strange behavior can occur if they're added out of order. But the
blotter 'txn' table is always a POSIXct index, so the comparison
fails if the user supplies data with a Date index.

Check if 'TxnDate' does not inherit from POSIXct (rather than checking
if it's time-based). If it's not POSIXct, check if it's Date and then
convert it to POSIXct with a UTC timezone. Otherwise just call
as.POSIXct() as before.

Fixes #51.