add infrastructure for managing flat periods in txnsim flat.to.reduced with replacement
This functionality is incomplete, see #48

We need to manage flat periods separately for the first layer of a potentially layered trade.  That means we need stats on the flat periods, so that we can do a statistically 'similar' sample.

These changes include some accounting for flat periods in txnsim.  It may be better to move some of this flat period work to perTradeStats.