add option to include flat periods in output of perTradeStats
flat periods are not really round turn trades, but they may be eifficently calculated from the posPL object which is manipulated by the perTradeStats function.  calculate them here if includeFlatPeriods=TRUE, and append to the end of perTradeStats output.

additionally, calculate duration and add to the output of perTradeStats, as it will be useful to multiple consumers, and is likewise easy to calculate here.

see #48