refactor txnsim increased.to.reduced replacement method
when sampling with replacement for increased.to.reduced, the complexity of the simulation is significantly increased f we want to maintain a semblance of the stylized facts of the original observed series.

This refactoring seeks to make the code easier to understand, and to better retain facts such as time in market and ratio of long to short trades.

This version of the code only applies the first full layer.  Additional layers still
need to be constructed and added to the function, as indicated by noted in the comments.

additionally, add documentation on nuances of sampling with replacement on increased.to.reduced.

see #48