fix incorrect last open trade for increased.to.reduced
Jasen Mackie noted (and fixed) that the last open trade Start time was incorrect.  We were taking the first increase from zero, not the last increase from zero.  In the case where there was more than one open trade at the end of the series, this would be incorrect.

see #48