add periodicity to perTradeStats, start on layering
periodicity of the posPL object is a stylized fact of the observed portfolio of transactions. By calculating and storing periodicity, we can make sure that replicates in txnsim preserve the periuodicity of the observed series.

in txnsim, we now utilize this periodicity to construct a time-based sequence with the same periodicity, bounded by the first trade timestamp and the end of the observed series.  we can then sample separately from this sdequence for each layer and side.

see #48