infrastructure to construct layers in txnsim
to properly layer trades, we need a roadmap of the state of the base layer, plus randomized places we can put additional layers.

this commit starts to put in that infrastructure, building the base layer, and constructing a sample list to work from for subsequent layers.   Only run this additional code if num_overlaps is greater than 1, it is not needed otherwise.

Additionally, we may want to test for percentages of positions at each level, and work the number of trades at the levels accordingly.

see #48