Fixes for Cum.PL, prorata, ts.prop and removing comments
Fix calculation bug for Cum.PL in the switch statement (once trades list is starting to be built) for flat.to.reduced. Also made absolute the calcs for prorata and ts.prop for flat.to.reduced and increased.to.reduced. Lastly, removed old code previously commented out in commit 58516a7.