Fixed Net.Trading.PL breakdown for increased.to.reduced
As opposed to using Period.Realized.PL from the posPL/trade object to compute realized PL we are now computing it directly from the txn price, avg cost and closing qty. This is necessitated by the fact that one closing txn can result in multiple round turns for the increased.to.reduced tradeDef.