fix portfolio names for txnsim, add helper function
txnsim needs to create a set of portfolios for the simulated transactions.  It follows that it needs to name these protfolios. we were constructing names on the fly, but not taking parameters of the call to txnsim into account.  I created a helper function to name these portfolios, and use this function in txnsim and plot.txnsim for now.  we may additionally want to allow special suffixes for these portfolios to allow for multiple calls to txnsim, but this solves the main problem I was having in that we couldn't call it both with and without replacement to see how they differed, because the portfolios would already exist.

Additionally, I undid some awful automatic code formatting done by RStudio reformat or reindent.  they would have made the code very difficult to maintain.