always updatePortf, add cumpl slot, add quantile method
The option to choose whether to updatePotf on the replicate portfolios didn't make anyt sense.  We need market data in order to generate the transactions, and we may not have that market data later when examining the results of the simulation.  So I removed the option.

Always calling updatePortf means that we can calculate cumpl here as well, so refactored to move that calculation out of plot.txnsim and into txnsim, creating a new cumpl slot.

refactor plot.txnsim to use the new cumpl slot rather than calculating separately

add quantile.txnsim function which uses the cumpl slot

update docs and NAMESPACE

bump version