scale Pos.Cost.Basis in perTradeStats by proportional position
In methods flat.to.reduced and increased.to.reduced for perTradeStats, we need to scale multiple calculations by the proportion of the total open position that is being included in this round turn.  We were not scaling Pos.Cost.Basis, which caused problems in Max.Notional.Cost, and all Pct measures.

Also, fix s/tmdf/smpdf/ typo in txnsim