changes to layering in txnsim
multiple changes while working on layering in txnsim.

the tmpdf object is used to help us layer trades on the first layer of long, short, and flat periods.  it is a data frame of round turn trades.  we were using a constructed 'pos' column, but this created incorrect positions.  Instead, using quantity to determine long/short.

Second, created a timesample function to sample from the time sequence across which the replicates are being layered. This eliminates code duplication for long and short samples, and makes the code easier to read.