extract convenience functions from txnsim
If you have a txnsim object, but don't have other objects that are created along the way, such as the portfolios, you may want to recreate these to do additional analysis.

This commit extracts and documents convenience functions from inside txnsim so that you could recreate the portfolios for further analysis.