use same layering for with/without replacement in flat.to.reduced and increased.to.reduced
The code for increased.to.reduced and flat.to reduced is now completely common across with and without replacement.  This changed the symsample.wr function and tradesample.wr and subsample functions to take a replacement argument.  The only practical changes were to set fudgefactor to 1 if replacement=FALSE and to sall sample() with replace=replacement in the subsample function.  All other code is now shared.

Additionally, minor changes to docs

bump version and Date prior to merge to master