add p-values for the test statistics
We have all the information to calculate p-values on the simulation test statistics, so add this to the function and output.

Additionally, add documentation on calculation and interpretation.

Further refactoring was done to remove duplicated code, include the backtest P&L in the appropriate places, and to name things a little more intelligently.