don't invert if we find the inverted pair in updatePosPL
if we find no exchange_rate information, updatePosPL will use 1 as the currency multiplier, which is correct.

if we find the correct exchange_rate, blotter will use it, which was also correct, or set invert=TRUE.

if we find the correct inverted exchange_rate, blotter was also setting invert=TRUE, which is wrong.

thanks to Talha Suhail for the report and reproducible example.