harmonize documentation for envir argument
harmonize the documentation between tradeStats and perTradeStats for taking an envir argument for the portfolio location.

I am a little worried that this should have a more descriptive name, like blotter_envir.

envir is used by getSymbols, so if tradeStats or perTradeStats are called in am embedded sequence of functions that also calls getSymbols using the envir argument, this will cause a name collision.  We might need to change this argument name in the future.

see #59