Add While loop counter and equality operator to IF condition 2 lines down
When layering trade durations onto the first layer in txnsim, we check to see if the sampled layering trade duration overlaps into another trade duration. If this happens, we check to see if the trade duration it overlaps is a similar side. If so, move onto the next layering long or short trade. If we instead overlap into a flat duration or an opposing side duration, we need to truncate.

Currently, we do not increment correctly down the list of first layer trades when looping >=2 times through the While loop, instead hard-coding an increment of 1 to the fixed first layer start timestamp. We need to add a counter to the While loop to increment appropriately for 2 or more loops.

Also, our While loop condition specifies only a > operator, but we need to account for the possibility that our first layer trade duration and our 2nd and subsequent layer trade durations could end on the same timestamp, and update the longdf and shortdf data frames accordingly.

This commit fixes these 2 somewhat related bugs.

see #61