Using try() As Opposed To is.null() For Checking If A Variable Exists
Inside txnstruct() we use is.null() to check for the existence of the maxlongpos and maxshortpos variables. This will not work if the variable does not exist. There are many ways to resolve this bug, so i will use a try() to check if calling missing(maxshortpos) returns an error.

see #64