Adding a condition for determining when to layer
In order to layer, we need to determine the index of the long and short duration dataframes (longdf & shortdf respectively) which take us over the duration of the first layer. If the last index before our target long and short durations is the final row in the long and short duration dataframes, we should not proceed to layering.

At present, we only check to see if the targetlongrow or targetshortrow == 0 when determining whether or not to proceed. I am adding another condition which checks that targetlongrow and targetshortrow > number of rows in longdf and shortdf before proceeding.

We could implement a solution higher up when targetlongrow and targetshortrow are first created, and render targetlongrow and/or targetshortrow == 0 if this is the case, but it feels more appropriate to add a condition to one which already exists for determining whether or not to proceed with layering.