Adding a check for NA values before subsetting any incomplete cases in cumpl
If there are no incomplete cases (such as NA values) as determined by the base R complete.cases() function, then we should not attempt to subset as this will result in a zero-width xts cumpl object which will throw a dimnames error when calling plot on the result of the call to txnsim().

In essence we will lose all the data from our cumpl xts object, which is the cumpl of each replicate.