Optimise construction of replicate transactions
Using a vectorised approach for the construction of the replicate transactions in txnsim.txns() we reduce the memory required to simulate n replicates. The improvement in performance is orders of magnitude, making the simulation of 1k replicates possible in the single digit minute time space (currently ranging from 20 mins to 15 hours depending on hardware) and making feasible simulations for a greater number of replicates.

see #65