Fix vectorised implementation for building txns, and fix bug related to incorrectly facing operator
The vectorised implementation for building our txns in txnsim.txns() uses findInterval() to assign timestamps and prices to our txns. When our original observed strategy is a strategy with layering, the vectorised solution implemented with commit c172724 introduces a bug, in that findInterval() will be called with an unordered input parameter (the layered trades with older timestamps will come after the more recent timestamps in the time series). To fix, we order the df before passing to findInterval().

Also, commit bf43488 introduced a bug in that layering was never possible with the operator the wrong way round. We switch the operator to be smaller than (<) nrow(shortdf) and nrow(longdf) to fix. This will be true when layering is required and will allow the algorithm to proceed to layering.

See #68
Fixes #68