reindent this code
pursuant to fixing #69, and from several other changes to this code, the indentation had gotten very inconsistent.

As I was investigating #69, that code formatting made the code harder to follow.  At the cost of messing up git blame, I'm reindenting this block of code to make it easier to work in in the future.

see #69