First commit to new branch 70_txnsim_layering
This commit has some significant changes to the way layering happens in the master branch. We layer until duration matches a predefined target, as opposed to only layering according to number of layers as determined by totaldur/calendardur.

The code is suboptimal, and while loops can run for a very very long time. Fixing this will be the subject of subsequent commits.

See #70

This commit also fixes #71