txnsim - adding a stylized fact and redoing layering process
When layering we need to know what the duration is between start timestamps of subsequent levels into an increasingly long or short position. We then add these durations randomly to randomly selected start timestamps of a first layer transaction whilst checkin we do not exceed duration of the first layer, nor the max pos. If we exceed maxpos with the subsequent layer, we truncate the txn qty so that we equal the max pos of the levelled into positions.