Remove pbo.cscv, consolidate into pbo, add vignette
It never really made sense to call the actual simulation function for CSCV as an internal function inside pbo. So we remove the internal function and add additional parameters to the pbo function definition.

We also update the documentation, and add a vignettes directory with a vignette replicating the Practical Application section in the PBO paper.

See #73