Update travis.yml so vignettes do not build
Build #185 errored with a message saying no output had been received in 10mins. We could configure travis to wait longer, see https://docs.travis-ci.com/user/common-build-problems/#Build-times-out-because-no-output-was-received. Looking at the travis yaml in quantstrat it appears we could also configure the travis build check to ignore vignettes. We use this approach in this commit.