Convert list portfolio to environment
Users are not able to access the portfolio environment. getPortfolio()
returns a list, to prevent the user from accidentally corrupting the
blotter portfolio. This causes the blotter portfolio to be saved as a
list, which means it can not be updated by reference.

Convert the user list portfolio to an environment, and copy attributes
(except for 'names', which is an error) to the environment.

Thanks to Thomas Huben for the report.