add aggregateArgs argument to pass args to aggregateFUN
aggregateFUN may take specific arguments.  Unfortunately, we can't just pass these in dots, because some of the other functions called by this function handle dots poorly.  We could try to find and remediate those, but this method has fewer possible unintended consequences.

It might have been better to use match.call and do.call here, but that would have been more complicated.  We can revisit that if this proves to be fragile.