Compress and re-save data to appease R CMD check
Ok, so i came across this commit message title from an earlier commit by @joshuaulrich in quanstrat and it gave me comfort that this was a real thing, a real requirement. So i found this useful question on SO (https://stackoverflow.com/questions/32605623/how-to-compress-saves-in-r-package-build), and ended up using devtools::use_data(ABC, compress="xz"). Thank you @hrbrmstr.