Extend benchTradePerf() (#95)
* Extend benchTradePerf()

Extension is in a number of directions:
- add 'TradeBench' benchmark
- compute rolling consistent quantities
- add summary and plot S3 methods
- remove multiple benchmark inputs

Documentation is updated accordingly.

* Remove unwanted example

The example is of course one of the ones in the documentation of the function, it was here only for a live testing session

* Minor performance threshold line adjustment

* Fix errors in docs

- document a param in summary.txnsPerf() method
- make a line in the examples .Rd compliant

* Docs files carpentry

- add S3method() and export() to NAMESPACE
- plot.txnsPerf() examples shifted in main function examples
- man/* updates

Note that other export() to NAMESPACE where generated. Keeping them as agreed.

* Change plot.txnsPerf() method main object name

According to "See section 'Generic functions and methods' in the 'Writing R extensions' manual" check suggestion

* Adjust RPM main plot average line appearence

* Conform RPM MktData subsetting

* Prevent p_avg benchmarking attemps when "new" market information not available

This applies to benchmark = 'VWAP', specifically the mechanism would be triggerend in the type = list(vwap = 'full') case. Also, it is impossible to know in advance the timestamps of p_avg and MktData that go under testing or which is their timing relation

* Add moving intervals to RPM

They are added to compute favorable, unfavorable and total market volume measures in time and as a result RPMs with respect to execution strategy prices.

Jasen's correction, thank you Jasen.

* Add initPortf() default initDate checker

Remove default initDate="1950-01-01" if portfolio object was initialized with this parameter

* Recycle past market data to benchmark later execution prices against

To guarantee that average execution prices are always benchmarked against last available market data, in a moving and irregural rolling manner. This appear to consistently apply to all possible timestamps mismatches

* Add RPM main plot bounds

RPM bounds are added to provide more an immediate visual context also to a non-technical audience.
Also, now a symmetric scaling of the y-axis should not break the grid under all plotting circumstances