Add DGP tests: Chi-Square and Kolmogorov-Smirnov GoF tests
The former is implemented as Kissell suggests (pp. 123--124), the latter is called from `stats` package. NAMESPACE and docs are updated accordingly. Also, new examples are added in docs.

One aspect which makes the function flow perhaphs less smooth and intuitive than needed is that for these tests Kissell suggests to use the arrival cost (a cost metric) as the preferred benchmark. This is in accordance to what suggested for the 'independent' approach, but in contrast suggestions for the 'paired' approach in which case performance metrics are considered (specifically, the VWAP).
There are clearly other reasons to keep the two approaches separate, but fix a common metric throughout may simplify some parts of the function; also taking into account that we deliberately kept out cost metrics from benchTradePerf().

Lastly, I shall stress some important details with respect to dgptest='ChiSq'. More specifically to how buckets are built. The way Kissell seems to suggest to use the standard-deviation based buckets building, namely {(-Inf,-3)*sd, [-3,-2.5]*sd, ..., [2.5, 3]*sd, (3, Inf)*sd}, is clearly not quite machine-friendly as it is. We could go through some fine tuning, but these considerations would neverthless be subordined to how findInterval() interprets numbers on close to the bounds and may quickly lead us to as deeply interesting as perhaps out-of-scope considerations. Although one may argue that chiBinStats will face a slight left-skewed bias, the test is ultimately based on their sum and thus results will be unbiased.
Therefore, to forgo cumbersome and perhpas unnecessary findInterval() splittings, the buckets have been litterally implemented as {(-Inf,-3]*sd, (-3,-2.5]*sd, ..., (2.5, 3]*sd, (3, Inf]*sd}. However and as an endnote on notation, the very last interval is printed as (3, Inf)*sd to avoid writing a mathematical abomination: Inf will never be reached anyways.