Replace deprecated indexClass() with tclass() in updatePosPL(), bump version
This should resolve a build failure for branch 54_algotca which includes an example for benchTradePerf() in which updatePortf() is called, which in turn calls updatePosPL() which calls indexClass(). Thanks to @joshuaulrich for helping identify the problem function in blotter, and for maintaining the awesome xts package. You can find his worthy sponsor page here, https://github.com/users/joshuaulrich/sponsorship.

see #54