Correct arrival price and rescale rolling quantities times
There are two interpretations on rolling variables. The first one is about computing them on period observed values and then shift the computed variable on the next day (with 'refIdx') as that is the first day they are "available". The second is to greedy compute them as soon as the market closed, then on the next day one should refer to the last available one (as would notmally happens in real-world circumstances when the market has not opened yet).
The former, however, is in contrast with comparison with the arrival price in that the last variables values would be comparable with an arrival price we haven't observed. Because of this we are now considering the latter