Add data-splitting via 'sessions' param
The 'MktData' original dataset is internally split using ISO time subsets, as suggested by @braverock. The basis remains a full trading day, to the widest of which 'sessions' defaults.

The scope of data-splitting is to augment the dataset, thus obtaining a larger number of data-points for our regression purposes.

All the rolling variables should be adapted to roll over sessions closings, in order to have meaningful comparisons and in particular to correctly evaluate the I-Star model equations in the end