Remove as.numeric() reconversions for 'MktValue'
The barest motivation of this reversion is about types coercion.

First of all, 'MktData' items may be originally coerced to "character" values, at least because of 'Reason'.

Coming from an all-numeric xts object, 'MktPrice' and 'MktQty' do not need any further conversion. This is already guaranteed in the function when the only non-numeric column in MktData items is 'Reason'.
Subsequently, taking advantage of the time-based nature of these objects, `do.call(rbind, *)` does exactly what intended, without requiring explicit handling. A further type conversion changes the obj class and breaks this flow.

When other non-numeric columns are present in 'MktData' items, the first conversion "fails" and thus the second type conversion is required to allow the above flow.

However, the former circumstance is the recommended one (see 'Details') and hence the one to support. Others non-numeric column are currently not required