Add 3D data grouping
Kissell reports that they perform data grouping before proceeding with the non-linear regression estimation. Here the grouping is optionally provided, to avoid shrinking excessively small given datasets or already too much 'horizon'/'sessions'-restricted variables vectors lengths.

The grouping is based on three variables: imbalance size, POV and volatility.
Buckets are built from sequences of user-provided bounds via 'targetIntervals', these bounds default to the author's suggested values. The minimum number of datapoints in a bucket so to be included in the sampled values is controlled with 'minDataPoints'.
Also note that, in a sense, a fourth dimension involved is on securities included in the anaysis.

Another feature provided is to average over the groups built, in order to "average away noise". This is optionally allowed, for the same reasons expressed above