Adjust and correct data grouping flow
In the case of grouped data only, the unlist() recursion is shorter than explicit flattenings.

For the groups averages is first of all introduced a correction, we average over groups values. Secondly, is put in place a simplified flattening with arguments similar as above. Note that the last na.omit(), although may appear repetitive, is indeed needed to account for excluded groups which would produce NaNs once averaged. Most further functions used internally handle these circumastances and results so far should be unchanged in this regard, specifically on regression. However, this last "cleaning" is done for sake of precision and memory saving.

If excluded groups are simply to be discarded a simpler, more intuitive and more efficient flow may be introduced