Introduce full model nonlinear regression
We estimate the full model, made of the instantaneous and market impact equations. Docs are updated accordingly.

Regression(s) in place so far was simply to have a way to reduce model complexity and possibly obtain some sort of hints on how the other pieces of workflow coded so far in terms of data splitting and data grouping would have influenced the analysis