Simplify data grouping 3D buckets building
Contrary to what stated before, buckets can be built taking advantage of the algorithm behind expand.grid. Or more precisely, cross them.

This is much simpler and avoids reinventing the wheel towards cleaner code.

The extraction is separated because there is still a for loop in place, better access a vector