Add check on MktData items required columns and refine docs
We require 'MktPrice' and 'MktQty' cols for each security in MktData.

A global check is put in place to avoid later time consuming stops or anyways unsatisfactory warnings on undetermined datasets.

Docs are updated accordingly and on top of this need is stressed that MktData items must be named. Apart from failures that it would otherwise raise, which can be worked around, it is important especially if one is working with hundreds of securities: this way we can provide helpful indications on which ones have to be corrected, if any