Add I-Star impact estimates and error measures
The model equations are added to estimate a given order impact costs. The order specifics will be provided through the newly added function parameter 'OrdData'.

Also, some other metrics such as the order "arrival cost" and error measures like the "cost error", "timing risk" and "z-score" are binded to the impact estimates output to provide immediate comparisons.

Further explanations are in the docs, which are updated accordingly