Solve secMktData and secMktSessions data types riddle
This solves the problem raised by @jaymon0703, https://github.com/braverock/blotter/pull/97#issuecomment-520415769 point 4.

Why converting to numeric from character yields an integer type in your case? This is because of the two natures a 'numeric' type has in R. Under the hood xts held a matrix type with integers in your case, no coercion to double as you do not have any. Applying a 'numeric' works, but a numeric of an integer is still an integer.

Clearly what will happen in practice depends at least on what 'MktPrice' users input, but this solves the problem in general.

Therefore no integer overflows anymore. Instead, a convenient coercion to NA will be practiced on unwanted "character"