Add sensitivity analysis via iStarSensitivity()
The new iStarSensitivity() is added to perform sensitivity analysis of model paramaters from the regression variables used in iStarPostTrade(). Regression variables datapoints can be full dataset datapoints, the groups or groups averages datapoints. Hence the new outstore[['Regression.Variables']] is added in iStarPostTrade() to later catch all these instances in iStarSensitivity().

It was separated from the main function in order to lessen the burden on it and to avoid confusions with the actual full model parameters estimation.

Explanations regarding the sensitivity analysis are provided in its docs. Among the other things, it should be stressed how nls() failures due to the actual model being estimated may cause the function to break: at the moment this is handled with tryCatch() returning NAs if errors are encountered, however we may want to provide more informative nls() failure messages as opposed to NAs (while still keeping NAs for parameters and RSS that cannot be provided)