Add iStarPostTrade backdoor to compute I-Star model with provided params
A user may have already estimated parameters for the model or anyways use parameters of his choice in the I-Star model equations, perhaps obtained using iStarSensitivity. In such instances the parameters estimation that we carry and its related pre-processing is not needed, instead have to be skipped.

The flow is adjusted accordingly, in particular the main 'outstore' object elements are now splitted by the task they result from. Explanations of the double nature of the 'OrdData' param are provided in the updated docs, together with their requirements.

The indentation is currently against style guidelines to avoid version control unwanted mismatches. This will be corrected once the review process is complete