Add cost curves S3 plot method
The iStarPostTrade output has been assigned 'iStarEst' class.

The added plot method allows using the output to generate cost curves, which have a qualitative purpose and can be useful to show how the model works. Explanations are provided in the added docs.

To remark is that iStarPostTrade() is internally reused to get consistent market impact estimations with respect to the ad hoc 'tmpOrdData' built