Adjust plot.iStarEst() and let 'sessions' default to NULL
The previous plot method was prone to confusions about the usage of 'OrdData'. This usage is now only internal and the user is simply required to input the parameters to use in cost curves generation.

Also, the method call to iStarPostTrade was stopped by the absence of a 'sessions' default value (which anyways was not used for those results). Therefore - and likely in a convenient manner for the main function itself - 'sessions' has now a new default construct