Fix overwriting bug in iStarSensitivity
Because of the previous structure the coefficients being written to be output object 'paramSens' where only from the models with the final fixed parameters values. The new commit aims to correct this circumstance.

Also, with the "if ladder" in place it appears there was no need of internal loops, 'parSeqsVals' should suffice.

Lastly, the new construct seems to be prone or at least suggestive to improvements that will likely follow in the future