Commit AlgoTCA folder, with R Journal article for GSoC 2019 Algo TCA project work
Potentially change folder name at a later date (to 'vignettes'?) once the document is complete, but when changing folder name to vignettes previously, changes in .Rmd were not rendered to the output Pdf. So we can leave the AlgoTCA folder named as such until the document is compelte, or leave it so named indefinitely.