Add Almgren-Chriss market impact model
The function 'acOptTxns' aims to add a first implementation of this model.

The model is summarized in the docs and links to the origianl sources are provided. Also, the implementation mimics the authors' original contribution and follows some of their assumptions and some of their extensions with respect to the baseline model, this is discussed in the docs as well