Adjust plots and code snippets
This commit is for sake of avoiding heavily trimmed or misplaced plots in the document.

Future revisions still need to address proper scales around and prevent margins to cover titles or axes, but in the meanwhile this will make it at least readable.

The same goes for the code snippets to show