Changes to 'longstartdiff' and 'shortstartdiff'
When determining a duration to add to a prior transaction start time for layering into a position, we were including distances between unrelated round turn trades. This commit fixes that and we now sample from durations inbetween layering start times which we observed in the original strategy. However, resulting sum of replicate durations are still less than the original strategy using our `bbands` demo.

At this point we will focus on documenting the process, which will aid future users and future 'me' and perhaps contribute to the start of a journal article.

We will ultimately need to solve the sampling process such that the sum of replicate durations more closely matches that observed in the original, but for now the current process should be documented.