reverse commit 9851c42
Utilizing lapply and do.call should avoid a loop, unecessary copies,
and constant memory reallocation on each loop iteration.

Mostly, that worked.

In some cases that I can't figure out, do.call was failing, and causing
a bunch of other things to fail and not get results.  So for now,
I am reverting commit 9851c42 to go back to the slower, but not buggy version.