Renaming Total.Net.Profit column to Net.Trading.PL
Although the column is currently named "Total.Net.Profit", the name in the output is "Net.Trading.PL", as the column name is being derived from last(Equity) which has the column name "Net.Trading.PL".
By renaming this column in the ret and tmpret data frames the column names in the output won't change when tradeStats is being called on a portfolio with transactions, however, the output will be different when tradeStats is run on a portfolio without transactions as the column name will change from "Total.Net.Profit" to "Net.Trading.PL". By renaming this will prevent the matching error which occurs when trying to rbind the tradeStats of the portfolios with transactions ("Net.Trading.PL") with the portfolios without ("Total.Net.Profit") when using the apply.paramset function.