Add a check for the class of index(prices)
When the index of the market data is a Date (as in the case of the macd demo) but the simulated portfolio timestamps are POSIX[c]t, the findInterval() function inside the txnsim.txns() function does not work properly. This commit should resolve that, and calling txnsim('macd', n=500) for example returns a beautiful simulations plot instead of only a single series of the original strategy.

Happy Sunday y'all